Fix fmap-duplicate performance issue
The fmap_duplicate function is used create a new fmap with a view into
an existing fmap. When the new view is a different size than the old
fmap, a new hash must be calculated for the duplicate fmap. However,
when the duplicated fmap is the same size as the original fmap, the hash
will be the same and there's no point recalculating.

The issue is apparent when scanning large EXE files because the hash was
being calculated at the beginning and end of the scan.

Digging into this issue revealed that hash calculations for fmaps were
also being performed at the wrong place. For scans of maps we use
fmap_duplicate() early in the process to apply the name API argument to
the duplicate fmap. Fixing the logic so we doing recalculate the hash
revealed that we never calculated hashes for fmap's created from buffers
in the first place, so that also had to be fixed be relocating where the
hash is calculated.

I also found that fmap_duplicate()'s offset argument used an off_t,
though it and all caller offsets are not allowed to be negative. This
was a bit of tangent to fix a bunch of off_t variables and paramters
that should've been size_t.

Added a couple unit tests to verify that making duplicate fmaps, and
duplicate-duplicate fmaps works as expected after the change.

Changed CLI_ISCONTAINED() and CLI_ISCONTAINED2() macros to cast to
size_t, because pointers and buffer sizes may not be negative, and these
two macros do not rely on substraction.

+++ libclamav/apm.c
@@ -53,7 +53,7 @@ int cli_scanapm(cli_ctx *ctx)
     struct apm_partition_info aptable, apentry;
     int ret = CL_CLEAN, detection = CL_CLEAN, old_school = 0;
     size_t sectorsize, maplen, partsize;
-    off_t pos = 0, partoff = 0;
+    size_t pos = 0, partoff = 0;
     unsigned i;
     uint32_t max_prtns = 0;
 
@@ -219,8 +219,8 @@ int cli_scanapm(cli_ctx *ctx)
         cli_dbgmsg("Type: %s\n", (char *)apentry.type);
         cli_dbgmsg("Signature: %x\n", apentry.signature);
         cli_dbgmsg("Partition Count: %u\n", apentry.numPartitions);
-        cli_dbgmsg("Blocks: [%u, +%u), ([%lu, +%lu))\n",
-                   apentry.pBlockStart, apentry.pBlockCount, (long unsigned)partoff, (long unsigned)partsize);
+        cli_dbgmsg("Blocks: [%u, +%u), ([%zu, +%zu))\n",
+                   apentry.pBlockStart, apentry.pBlockCount, partoff, partsize);
 
         /* send the partition to cli_magic_scan_nested_fmap_type */
         ret = cli_magic_scan_nested_fmap_type(*ctx->fmap, partoff, partsize, ctx, CL_TYPE_PART_ANY, (const char *)apentry.name);
@@ -245,7 +245,7 @@ static int apm_partition_intersection(cli_ctx *ctx, struct apm_partition_info *a
     struct apm_partition_info apentry;
     unsigned i, pitxn;
     int ret = CL_CLEAN, tmp = CL_CLEAN;
-    off_t pos;
+    size_t pos;
     uint32_t max_prtns = 0;
     int virus_found    = 0;
